+++ android/src/main/java/org/inaturalist/inatcamera/classifier/Prediction.java
@@ -3,7 +3,7 @@ package org.inaturalist.inatcamera.classifier;
 public class Prediction {
     public Node node;
     public Double probability;
-    public Integer rank;
+    public Float rank;
 
     public Prediction(Node n, double p) {
         node = n;
